modularizes rest_api test suite
