+++ tests/sources/rest_api/configurations/test_custom_auth_config.py
